feat(sdk): Export `RoomListEntry` into its own module.
